Move BlockError test to errors mod; add SecretBlockError tost
